pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "reachy-mini-companion"
version = "0.1.0"
description = "AI-powered desk companion for Reachy Mini robot"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
authors = [
    {name = "Jan Moons", email = "jan@example.com"}
]
dependencies = [
    "reachy-mini>=1.2.4",
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
    "google-genai>=1.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "ruff>=0.1.0",
]

[project.entry-points.reachy_mini_apps]
reachy-mini-companion = "reachy_mini_companion.main:ReachyMiniCompanion"

[tool.hatch.build.targets.wheel]
packages = ["reachy_mini_companion"]
